refactor: extract ChatInterface logic into modular components
BREAKING DOWN LARGE COMPONENT:
- Extract workflow detection logic into useWorkflowDetection hook
- Extract model status management into useModelStatus hook
- Extract CopilotKit actions into separate copilotActions.ts file
- Improve type safety with proper TypeScript interfaces

NEW MODULES CREATED:
- src/hooks/useWorkflowDetection.ts - Smart workflow opportunity detection with confidence scoring
- src/hooks/useModelStatus.ts - AI model initialization progress tracking and status management
- src/lib/copilotActions.ts - All CopilotKit action definitions with transitionToWorkflow helper

BENEFITS:
✅ Reduced ChatInterface from 628 lines to ~190 lines (70% reduction)
✅ Improved separation of concerns and single responsibility principle
✅ Enhanced reusability - hooks can be used in other components
✅ Better testability - each module can be tested independently
✅ Cleaner code organization following project architecture guidelines
✅ Proper TypeScript types replacing 'any' with specific interfaces
✅ Maintained all existing functionality while improving maintainability

TECHNICAL IMPROVEMENTS:
- Added proper TypeScript types (AppRouterInstance, WorkflowSuggestion)
- Fixed journey type mapping (blog-creation, email-campaign, social-media)
- Preserved all CopilotKit integration and Bilan analytics tracking
- Maintained workflow suggestion banner and context preservation
- Kept model status checking and streaming progress indicators
