+++ src/components/ChatInterface.tsx
@@ -3,7 +3,6 @@
 import { useState, useEffect } from 'react'
 import { CopilotKit } from '@copilotkit/react-core'
 import { CopilotChat } from '@copilotkit/react-ui'
-import { useCopilotAction } from '@copilotkit/react-core'
 import {
   Container,
   Stack,
@@ -30,10 +29,11 @@ import {
   startConversation, 
   endConversation, 
   initializeBilan,
-  createUserId,
-  startJourney
+  createUserId
 } from '../lib/bilan'
 import { useRouter } from 'next/navigation'
+import { useWorkflowDetection, useModelStatus } from '../hooks'
+import { useCopilotActions, transitionToWorkflow } from '../lib/copilotActions'
 
 /**
  * Chat interface component with CopilotKit integration and comprehensive Bilan tracking
@@ -48,16 +48,12 @@ import { useRouter } from 'next/navigation'
  * - Intelligent workflow detection and smooth context transitions
  * 
  * @component
- * @returns {JSX.Element} Full-featured chat interface with analytics tracking
  */
 export function ChatInterface() {
   const [conversationId, setConversationId] = useState<string>('')
   const [isLoading, setIsLoading] = useState(true)
   const [error, setError] = useState<string>('')
   const [userId] = useState(() => createUserId(`user_${Date.now()}_${Math.random().toString(36).slice(2, 11)}`))
-  const [modelProgress, setModelProgress] = useState<number>(0)
-  const [modelStatus, setModelStatus] = useState<'idle' | 'loading' | 'ready' | 'error'>('idle')
-  const [statusMessage, setStatusMessage] = useState<string>('')
   const [workflowSuggestion, setWorkflowSuggestion] = useState<{
     type: 'blog' | 'email' | 'social'
     reason: string
@@ -66,335 +62,18 @@ export function ChatInterface() {
   } | null>(null)
   const router = useRouter()
 
-  // Check AI model status and initialization progress
-  const checkModelStatus = async () => {
-    try {
-      const response = await fetch('/api/copilot-kit', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          messages: [{ role: 'user', content: 'status check' }],
-          max_tokens: 1
-        })
-      })
+  // Extract model status management to custom hook
+  const { modelProgress, modelStatus, statusMessage, checkModelStatus } = useModelStatus()
+  
+  // Extract workflow detection logic to custom hook
+  const { detectWorkflowOpportunity } = useWorkflowDetection()
 
-      const data = await response.json()
-      
-      if (data.status === 'initializing') {
-        setModelStatus('loading')
-        setModelProgress(data.progress || 0)
-        setStatusMessage(data.message || 'Loading AI model...')
-        
-        // Continue checking until ready
-        setTimeout(checkModelStatus, 1000)
-      } else if (data.error) {
-        setModelStatus('error')
-        setStatusMessage(data.error.message || 'Model initialization failed')
-      } else {
-        setModelStatus('ready')
-        setModelProgress(100)
-        setStatusMessage('AI model ready')
-      }
-    } catch (error) {
-      setModelStatus('error')
-      setStatusMessage(`Failed to check model status: ${error instanceof Error ? error.message : 'Unknown error'}`)
-    }
-  }
-
-  // Intelligent workflow detection based on user input
-  const detectWorkflowOpportunity = (message: string): {
-    type: 'blog' | 'email' | 'social'
-    reason: string
-    context: string
-    confidence: number
-  } | null => {
-    const lowerMessage = message.toLowerCase()
-    
-    // Blog detection patterns
-    const blogPatterns = [
-      'write a blog', 'blog post', 'article about', 'tutorial on', 
-      'guide to', 'how to', 'step by step', 'comprehensive overview',
-      'deep dive', 'analysis of', 'case study', 'review of'
-    ]
-    
-    // Email detection patterns  
-    const emailPatterns = [
-      'write an email', 'email to', 'professional email', 'send email',
-      'follow up email', 'announcement email', 'newsletter', 'outreach',
-      'client email', 'team update', 'business email'
-    ]
-    
-    // Social media detection patterns
-    const socialPatterns = [
-      'social media post', 'tweet', 'linkedin post', 'instagram caption',
-      'facebook post', 'social content', 'viral post', 'engagement post',
-      'announcement post', 'behind the scenes'
-    ]
-
-    // Check for blog opportunities
-    for (const pattern of blogPatterns) {
-      if (lowerMessage.includes(pattern)) {
-        return {
-          type: 'blog',
-          reason: `Detected request for long-form content: "${pattern}"`,
-          context: message,
-          confidence: 0.85
-        }
-      }
-    }
-
-    // Check for email opportunities
-    for (const pattern of emailPatterns) {
-      if (lowerMessage.includes(pattern)) {
-        return {
-          type: 'email',
-          reason: `Detected email communication need: "${pattern}"`,
-          context: message,
-          confidence: 0.9
-        }
-      }
-    }
-
-    // Check for social media opportunities
-    for (const pattern of socialPatterns) {
-      if (lowerMessage.includes(pattern)) {
-        return {
-          type: 'social',
-          reason: `Detected social media content request: "${pattern}"`,
-          context: message,
-          confidence: 0.8
-        }
-      }
-    }
-
-    return null
-  }
-
-  // Enhanced workflow transition with context preservation
-  const transitionToWorkflow = async (
-    workflowType: 'blog' | 'email' | 'social',
-    context: string,
-    preserveConversation = true
-  ) => {
-    try {
-      // Start journey tracking for workflow transition
-      const journeyId = await startJourney(`${workflowType}-creation` as any, {
-        topic: context,
-        userBrief: context,
-        contentType: workflowType,
-        initialConversationId: preserveConversation ? conversationId : undefined,
-        transitionSource: 'chat-interface',
-        transitionReason: workflowSuggestion?.reason
-      })
-
-      // Prepare URL parameters with context
-      const params = new URLSearchParams({
-        type: workflowType,
-        context: context.substring(0, 500), // Limit context length for URL
-        fromChat: 'true',
-        journeyId,
-        ...(conversationId && preserveConversation && { conversationId })
-      })
-
-      // Navigate to workflow with preserved context
-      router.push(`/create?${params.toString()}`)
-      
-      // Clear workflow suggestion
-      setWorkflowSuggestion(null)
-    } catch (error) {
-      console.error('Failed to transition to workflow:', error)
-    }
-  }
-
-  // Custom CopilotKit Actions for Content Creation Tools
-  useCopilotAction({
-    name: 'startBlogWorkflow',
-    description: 'Start a structured blog post creation workflow with step-by-step guidance',
-    parameters: [
-      {
-        name: 'topic',
-        type: 'string',
-        description: 'The topic or title for the blog post',
-        required: true
-      },
-      {
-        name: 'audience',
-        type: 'string', 
-        description: 'Target audience for the blog post',
-        required: false
-      }
-    ],
-    handler: async ({ topic, audience }) => {
-      await transitionToWorkflow('blog', `Topic: ${topic}${audience ? ` | Audience: ${audience}` : ''}`)
-      return `Starting blog workflow for "${topic}". Redirecting to structured creation process with chat context preserved...`
-    }
-  })
-
-  useCopilotAction({
-    name: 'startEmailWorkflow',
-    description: 'Start a structured email creation workflow for professional communications',
-    parameters: [
-      {
-        name: 'purpose',
-        type: 'string',
-        description: 'The purpose or goal of the email',
-        required: true
-      },
-      {
-        name: 'recipient',
-        type: 'string',
-        description: 'Who the email is for (e.g., clients, team, customers)',
-        required: false
-      }
-    ],
-    handler: async ({ purpose, recipient }) => {
-      await transitionToWorkflow('email', `Purpose: ${purpose}${recipient ? ` | Recipient: ${recipient}` : ''}`)
-      return `Starting email workflow for "${purpose}". Redirecting to structured creation process with chat context preserved...`
-    }
-  })
-
-  useCopilotAction({
-    name: 'startSocialWorkflow', 
-    description: 'Start a structured social media content creation workflow',
-    parameters: [
-      {
-        name: 'platform',
-        type: 'string',
-        description: 'Target social media platform (Twitter, LinkedIn, Instagram, etc.)',
-        required: false
-      },
-      {
-        name: 'goal',
-        type: 'string',
-        description: 'Goal of the social media post (engagement, awareness, promotion, etc.)',
-        required: true
-      }
-    ],
-    handler: async ({ platform, goal }) => {
-      await transitionToWorkflow('social', `Goal: ${goal}${platform ? ` | Platform: ${platform}` : ''}`)
-      return `Starting social media workflow for ${platform ? `${platform} ` : ''}with goal: "${goal}". Redirecting to structured creation process with chat context preserved...`
-    }
-  })
-
-  useCopilotAction({
-    name: 'improveText',
-    description: 'Improve existing text by making it clearer, more engaging, or fixing grammar',
-    parameters: [
-      {
-        name: 'text',
-        type: 'string',
-        description: 'The text to improve',
-        required: true
-      },
-      {
-        name: 'improvementType',
-        type: 'string',
-        description: 'Type of improvement: clarity, engagement, grammar, or conciseness',
-        required: false
-      }
-    ],
-    handler: async ({ text, improvementType = 'general' }) => {
-      // This could be enhanced to use different prompts based on improvement type
-      return `Here's the improved version of your text:
-
-**Original:**
-${text}
-
-**Improved (${improvementType}):**
-I'll help you improve this text. Let me analyze it and provide suggestions for better ${improvementType}.
-
-*Note: For more sophisticated text improvement, consider using the structured workflows for specific content types.*`
-    }
-  })
-
-  useCopilotAction({
-    name: 'generateOutline',
-    description: 'Generate a structured outline for any type of content',
-    parameters: [
-      {
-        name: 'topic',
-        type: 'string', 
-        description: 'The topic for the outline',
-        required: true
-      },
-      {
-        name: 'contentType',  
-        type: 'string',
-        description: 'Type of content: blog, article, presentation, email, etc.',
-        required: false
-      },
-      {
-        name: 'length',
-        type: 'string',
-        description: 'Desired length: short, medium, long, or detailed',
-        required: false
-      }
-    ],
-    handler: async ({ topic, contentType = 'general', length = 'medium' }) => {
-      return `# Content Outline: ${topic}
-
-**Type:** ${contentType} | **Length:** ${length}
-
-## I. Introduction
-- Hook/Opening statement
-- Background context
-- Main thesis or purpose
-
-## II. Main Content
-- Key point 1
-- Key point 2  
-- Key point 3
-- Supporting details and examples
-
-## III. Conclusion
-- Summary of main points
-- Call to action or next steps
-- Closing thought
-
-*This is a basic outline. For more detailed, step-by-step content creation with analytics tracking, use the structured workflows available in the Workflows tab.*`
-    }
-  })
-
-  useCopilotAction({
-    name: 'analyzeContent',
-    description: 'Analyze content for readability, tone, structure, and provide improvement suggestions',
-    parameters: [
-      {
-        name: 'content',
-        type: 'string',
-        description: 'The content to analyze',
-        required: true
-      },
-      {
-        name: 'analysisType',
-        type: 'string', 
-        description: 'Focus area: readability, tone, structure, or comprehensive',
-        required: false
-      }
-    ],
-    handler: async ({ content, analysisType = 'comprehensive' }) => {
-      const wordCount = content.split(' ').length
-      const sentences = content.split(/[.!?]+/).length - 1
-      const avgWordsPerSentence = sentences > 0 ? Math.round(wordCount / sentences) : 0
-
-      return `# Content Analysis (${analysisType})
-
-**Content Statistics:**
-- Word count: ${wordCount}
-- Sentences: ${sentences}
-- Average words per sentence: ${avgWordsPerSentence}
-
-**Quick Assessment:**
-- Readability: ${avgWordsPerSentence < 20 ? 'Good' : 'Could be improved (long sentences)'}
-- Length: ${wordCount < 100 ? 'Short' : wordCount < 500 ? 'Medium' : 'Long'}
-
-**Recommendations:**
-1. ${avgWordsPerSentence > 25 ? 'Consider breaking up long sentences for better readability' : 'Sentence length is appropriate'}
-2. Check for consistent tone throughout
-3. Ensure clear structure with logical flow
-
-*For detailed content optimization with A/B testing insights, use our structured workflows.*`
-    }
+  // Initialize CopilotKit actions
+  useCopilotActions({
+    conversationId,
+    workflowSuggestion,
+    router,
+    setWorkflowSuggestion
   })
 
   // Initialize Bilan and start conversation
@@ -404,21 +83,18 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
     const initializeChat = async () => {
       try {
         setIsLoading(true)
-        setModelStatus('loading')
         
         // Start model status checking
         checkModelStatus()
         
-        // Initialize Bilan SDK
         await initializeBilan(userId)
         
         if (!mounted) return
-
-        // Start conversation for chat mode
+        
         const convId = await startConversation({
           topic: 'general-chat',
           userIntent: 'open-conversation',
-          contentType: 'social' // Using social as it's most conversational
+          contentType: 'social'
         })
         
         if (mounted) {
@@ -430,31 +106,24 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
         if (mounted) {
           setError('Failed to initialize chat. Please refresh the page.')
           setIsLoading(false)
-          setModelStatus('error')
         }
       }
     }
 
     initializeChat()
 
-    // Cleanup: end conversation when component unmounts
     return () => {
       mounted = false
       if (conversationId) {
         endConversation(conversationId, 'completed', {
-          satisfactionScore: 5, // Default positive score for completed chats
+          satisfactionScore: 5,
           outcome: 'natural-completion'
         }).catch(() => {
           // Ignore cleanup errors
         })
       }
     }
-  }, [userId, conversationId])
-  
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-
-  // Note: Turn tracking and voting will be implemented in the next commit
-  // This requires deeper integration with CopilotKit's message lifecycle
+  }, [userId, conversationId, checkModelStatus]) // eslint-disable-line react-hooks/exhaustive-deps
 
   if (error) {
     return (
@@ -482,9 +151,7 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
               </ThemeIcon>
               <div style={{ flex: 1 }}>
                 <Group gap="sm" align="center">
-                  <Text size="lg" fw={600}>
-                    AI Writing Assistant
-                  </Text>
+                  <Text size="lg" fw={600}>AI Writing Assistant</Text>
                   <Badge 
                     size="sm" 
                     color={modelStatus === 'ready' ? 'green' : modelStatus === 'loading' ? 'yellow' : 'red'}
@@ -498,9 +165,7 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
                     </Group>
                   </Badge>
                 </Group>
-                <Text size="sm" c="dimmed">
-                  Powered by WebLLM • Analytics by Bilan • Smart Workflow Detection
-                </Text>
+                <Text size="sm" c="dimmed">Powered by WebLLM • Analytics by Bilan • Smart Workflow Detection</Text>
                 {modelStatus === 'loading' && (
                   <div style={{ marginTop: 4 }}>
                     <Progress value={modelProgress} size="xs" />
@@ -521,9 +186,7 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
                       <IconBulb size={14} />
                     </ThemeIcon>
                     <div>
-                      <Text size="sm" fw={500}>
-                        Workflow Suggestion
-                      </Text>
+                      <Text size="sm" fw={500}>Workflow Suggestion</Text>
                       <Text size="xs" c="dimmed">
                         {workflowSuggestion.reason} (Confidence: {Math.round(workflowSuggestion.confidence * 100)}%)
                       </Text>
@@ -534,7 +197,14 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
                       size="xs"
                       variant="filled"
                       rightSection={<IconArrowRight size={12} />}
-                      onClick={() => transitionToWorkflow(workflowSuggestion.type, workflowSuggestion.context)}
+                      onClick={() => transitionToWorkflow(
+                        workflowSuggestion.type, 
+                        workflowSuggestion.context,
+                        conversationId,
+                        workflowSuggestion,
+                        router,
+                        setWorkflowSuggestion
+                      )}
                     >
                       Start {workflowSuggestion.type} workflow
                     </Button>
@@ -550,7 +220,7 @@ I'll help you improve this text. Let me analyze it and provide suggestions for b
               </Paper>
             )}
 
-            {/* CopilotChat Component with Enhanced Instructions */}
+            {/* CopilotChat Interface */}
             <div style={{ flex: 1, minHeight: 0 }}>
               <CopilotChat
                 labels={{
@@ -598,25 +268,15 @@ What would you like to work on today?`,
 - Always be conversational and explain your reasoning
 
 Remember: You can detect workflow needs and suggest smooth transitions while preserving context!`}
-                onInProgress={(inProgress) => {
-                  // Update progress indicator when AI is generating
-                  if (inProgress) {
-                    setStatusMessage('Generating response...')
-                  } else {
-                    setStatusMessage('AI model ready')
-                  }
+                onInProgress={(_inProgress) => {
+                  // Optional: Could update status message here
                 }}
                 onSubmitMessage={async (message: string) => {
-                  // Detect workflow opportunities in user input
                   const workflowOpp = detectWorkflowOpportunity(message)
                   if (workflowOpp && workflowOpp.confidence > 0.7) {
                     setWorkflowSuggestion(workflowOpp)
                   }
-                  
-                  setStatusMessage('Processing request...')
                 }}
-                // Note: Voting integration will be implemented in next commit
-                // CopilotKit's voting API may need different approach
                 className="h-full"
               />
             </div>
