+++ src/hooks/index.ts
@@ -1,3 +1,7 @@
 // Navigation hooks
 export { useNavigation } from './useNavigation'
-export type { NavigationState } from './useNavigation' 
+export type { NavigationState } from './useNavigation'
+
+// Chat interface hooks
+export { useWorkflowDetection } from './useWorkflowDetection'
+export { useModelStatus } from './useModelStatus'
